refactor(docx): dissolve DrawingML host package
