+++ packages/docx/package.json
@@ -209,7 +209,6 @@
     "build:wasm": "bun ../../scripts/build-docx-wasm.ts"
   },
   "devDependencies": {
-    "@betteroffice/drawingml": "workspace:*",
     "tsup": "^8.0.1",
     "typescript": "^6.0.3",
     "@types/bun": "^1.0.0",
