+++ packages/docx/src/types/drawingml.ts
@@ -1,0 +1,304 @@
+/** DrawingML value contracts retained by the DOCX host. */
+
+export type ThemeColorSlot =
+  | 'dk1'
+  | 'lt1'
+  | 'dk2'
+  | 'lt2'
+  | 'accent1'
+  | 'accent2'
+  | 'accent3'
+  | 'accent4'
+  | 'accent5'
+  | 'accent6'
+  | 'hlink'
+  | 'folHlink'
+  | 'background1'
+  | 'text1'
+  | 'background2'
+  | 'text2';
+
+export interface ColorValue {
+  rgb?: string;
+  themeColor?: ThemeColorSlot;
+  themeTint?: string;
+  themeShade?: string;
+  auto?: boolean;
+}
+
+export interface ThemeColorScheme {
+  dk1?: string;
+  lt1?: string;
+  dk2?: string;
+  lt2?: string;
+  accent1?: string;
+  accent2?: string;
+  accent3?: string;
+  accent4?: string;
+  accent5?: string;
+  accent6?: string;
+  hlink?: string;
+  folHlink?: string;
+}
+
+export interface ThemeFont {
+  latin?: string;
+  ea?: string;
+  cs?: string;
+  fonts?: Record<string, string>;
+}
+
+export interface ThemeFontScheme {
+  majorFont?: ThemeFont;
+  minorFont?: ThemeFont;
+}
+
+export interface Theme {
+  name?: string;
+  colorScheme?: ThemeColorScheme;
+  fontScheme?: ThemeFontScheme;
+  formatScheme?: {
+    name?: string;
+  };
+}
+
+export type ShapeType =
+  | 'rect'
+  | 'roundRect'
+  | 'ellipse'
+  | 'triangle'
+  | 'rtTriangle'
+  | 'parallelogram'
+  | 'trapezoid'
+  | 'pentagon'
+  | 'hexagon'
+  | 'heptagon'
+  | 'octagon'
+  | 'decagon'
+  | 'dodecagon'
+  | 'star4'
+  | 'star5'
+  | 'star6'
+  | 'star7'
+  | 'star8'
+  | 'star10'
+  | 'star12'
+  | 'star16'
+  | 'star24'
+  | 'star32'
+  | 'line'
+  | 'straightConnector1'
+  | 'bentConnector2'
+  | 'bentConnector3'
+  | 'bentConnector4'
+  | 'bentConnector5'
+  | 'curvedConnector2'
+  | 'curvedConnector3'
+  | 'curvedConnector4'
+  | 'curvedConnector5'
+  | 'rightArrow'
+  | 'leftArrow'
+  | 'upArrow'
+  | 'downArrow'
+  | 'leftRightArrow'
+  | 'upDownArrow'
+  | 'quadArrow'
+  | 'leftRightUpArrow'
+  | 'bentArrow'
+  | 'uturnArrow'
+  | 'leftUpArrow'
+  | 'bentUpArrow'
+  | 'curvedRightArrow'
+  | 'curvedLeftArrow'
+  | 'curvedUpArrow'
+  | 'curvedDownArrow'
+  | 'stripedRightArrow'
+  | 'notchedRightArrow'
+  | 'homePlate'
+  | 'chevron'
+  | 'rightArrowCallout'
+  | 'downArrowCallout'
+  | 'leftArrowCallout'
+  | 'upArrowCallout'
+  | 'leftRightArrowCallout'
+  | 'quadArrowCallout'
+  | 'circularArrow'
+  | 'flowChartProcess'
+  | 'flowChartAlternateProcess'
+  | 'flowChartDecision'
+  | 'flowChartInputOutput'
+  | 'flowChartPredefinedProcess'
+  | 'flowChartInternalStorage'
+  | 'flowChartDocument'
+  | 'flowChartMultidocument'
+  | 'flowChartTerminator'
+  | 'flowChartPreparation'
+  | 'flowChartManualInput'
+  | 'flowChartManualOperation'
+  | 'flowChartConnector'
+  | 'flowChartOffpageConnector'
+  | 'flowChartPunchedCard'
+  | 'flowChartPunchedTape'
+  | 'flowChartSummingJunction'
+  | 'flowChartOr'
+  | 'flowChartCollate'
+  | 'flowChartSort'
+  | 'flowChartExtract'
+  | 'flowChartMerge'
+  | 'flowChartOnlineStorage'
+  | 'flowChartDelay'
+  | 'flowChartMagneticTape'
+  | 'flowChartMagneticDisk'
+  | 'flowChartMagneticDrum'
+  | 'flowChartDisplay'
+  | 'wedgeRectCallout'
+  | 'wedgeRoundRectCallout'
+  | 'wedgeEllipseCallout'
+  | 'cloudCallout'
+  | 'borderCallout1'
+  | 'borderCallout2'
+  | 'borderCallout3'
+  | 'accentCallout1'
+  | 'accentCallout2'
+  | 'accentCallout3'
+  | 'callout1'
+  | 'callout2'
+  | 'callout3'
+  | 'accentBorderCallout1'
+  | 'accentBorderCallout2'
+  | 'accentBorderCallout3'
+  | 'actionButtonBlank'
+  | 'actionButtonHome'
+  | 'actionButtonHelp'
+  | 'actionButtonInformation'
+  | 'actionButtonBackPrevious'
+  | 'actionButtonForwardNext'
+  | 'actionButtonBeginning'
+  | 'actionButtonEnd'
+  | 'actionButtonReturn'
+  | 'actionButtonDocument'
+  | 'actionButtonSound'
+  | 'actionButtonMovie'
+  | 'irregularSeal1'
+  | 'irregularSeal2'
+  | 'frame'
+  | 'halfFrame'
+  | 'corner'
+  | 'diagStripe'
+  | 'chord'
+  | 'arc'
+  | 'bracketPair'
+  | 'bracePair'
+  | 'leftBracket'
+  | 'rightBracket'
+  | 'leftBrace'
+  | 'rightBrace'
+  | 'can'
+  | 'cube'
+  | 'bevel'
+  | 'donut'
+  | 'noSmoking'
+  | 'blockArc'
+  | 'foldedCorner'
+  | 'smileyFace'
+  | 'heart'
+  | 'lightningBolt'
+  | 'sun'
+  | 'moon'
+  | 'cloud'
+  | 'snip1Rect'
+  | 'snip2SameRect'
+  | 'snip2DiagRect'
+  | 'snipRoundRect'
+  | 'round1Rect'
+  | 'round2SameRect'
+  | 'round2DiagRect'
+  | 'plaque'
+  | 'teardrop'
+  | 'mathPlus'
+  | 'mathMinus'
+  | 'mathMultiply'
+  | 'mathDivide'
+  | 'mathEqual'
+  | 'mathNotEqual'
+  | 'gear6'
+  | 'gear9'
+  | 'funnel'
+  | 'pieWedge'
+  | 'pie'
+  | 'leftCircularArrow'
+  | 'leftRightCircularArrow'
+  | 'swooshArrow'
+  | 'textBox';
+
+export interface ShapeFill {
+  type: 'none' | 'solid' | 'gradient' | 'pattern' | 'picture';
+  color?: ColorValue;
+  gradient?: {
+    type: 'linear' | 'radial' | 'rectangular' | 'path';
+    angle?: number;
+    stops: Array<{
+      position: number;
+      color: ColorValue;
+    }>;
+  };
+}
+
+export interface ShapeOutline {
+  width?: number;
+  color?: ColorValue;
+  style?:
+    | 'solid'
+    | 'dot'
+    | 'dash'
+    | 'lgDash'
+    | 'dashDot'
+    | 'lgDashDot'
+    | 'lgDashDotDot'
+    | 'sysDot'
+    | 'sysDash'
+    | 'sysDashDot'
+    | 'sysDashDotDot';
+  cap?: 'flat' | 'round' | 'square';
+  join?: 'bevel' | 'miter' | 'round';
+  headEnd?: {
+    type: 'none' | 'triangle' | 'stealth' | 'diamond' | 'oval' | 'arrow';
+    width?: 'sm' | 'med' | 'lg';
+    length?: 'sm' | 'med' | 'lg';
+  };
+  tailEnd?: {
+    type: 'none' | 'triangle' | 'stealth' | 'diamond' | 'oval' | 'arrow';
+    width?: 'sm' | 'med' | 'lg';
+    length?: 'sm' | 'med' | 'lg';
+  };
+}
+
+export interface ShapeTextBody<TContent = unknown> {
+  vertical?: boolean;
+  rotation?: number;
+  anchor?: 'top' | 'middle' | 'bottom' | 'distributed' | 'justified';
+  anchorCenter?: boolean;
+  autoFit?: 'none' | 'normal' | 'shape';
+  margins?: {
+    top?: number;
+    bottom?: number;
+    left?: number;
+    right?: number;
+  };
+  content: TContent[];
+}
+
+export type PresetGeometryPathCommand =
+  | { type: 'move'; x: number; y: number }
+  | { type: 'line'; x: number; y: number }
+  | { type: 'quad'; cpx: number; cpy: number; x: number; y: number }
+  | {
+      type: 'cubic';
+      cp1x: number;
+      cp1y: number;
+      cp2x: number;
+      cp2y: number;
+      x: number;
+      y: number;
+    }
+  | { type: 'close' };
