--- packages/drawingml/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "@betteroffice/drawingml",
-  "version": "0.0.1",
-  "private": true,
-  "description": "Shared DrawingML (a: namespace) primitives for OOXML formats - theme parsing, color math, shape fill/outline/transform parsing, and serialization.",
-  "sideEffects": false,
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.mjs",
-      "require": "./dist/index.js"
-    },
-    "./units": {
-      "types": "./dist/units.d.ts",
-      "import": "./dist/units.mjs",
-      "require": "./dist/units.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "typesVersions": {
-    "*": {
-      "units": [
-        "./dist/units.d.ts"
-      ]
-    }
-  },
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsup",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "tsup": "^8.0.1",
-    "typescript": "^6.0.3"
-  },
-  "keywords": [
-    "ooxml",
-    "drawingml",
-    "docx",
-    "pptx",
-    "xlsx",
-    "theme",
-    "colors"
-  ],
-  "author": "OpenOOXML",
-  "license": "Apache-2.0",
-  "homepage": "https://betteroffice.dev/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/openooxml/betteroffice.git",
-    "directory": "packages/drawingml"
-  }
-}
