--- packages/drawingml/temp/units.api.md
@@ -1,25 +1,0 @@
-## API Report File for "@betteroffice/drawingml"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-// @public
-export const EMUS_PER_INCH = 914400;
-
-// @public
-export function emuToPixels(emu: number | undefined | null): number;
-
-// @public
-export const PIXELS_PER_INCH = 96;
-
-// @public
-export function pixelsToEmu(px: number): number;
-
-// @public
-export const POINTS_PER_INCH = 72;
-
-// @public
-export function pointsToPixels(points: number): number;
-
-```
