--- packages/drawingml/tsup.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: {
-    index: 'src/index.ts',
-    units: 'src/units.ts',
-  },
-  format: ['cjs', 'esm'],
-  dts: true,
-  splitting: false,
-  sourcemap: false,
-  clean: true,
-  minify: false,
-});
